redo the actual changes
the first commit in master had to be reverted, since the tests were still in a temporary modified state. A new branch (textOptions) was then created on the basis of the meanwhile reverted commit. In this branch, the tests were cleaned up and the branch finally got merged with tha master. However, the actual changes, which were reverted in the master, were not merged, despiute the fact they were present in the brach. Finally, I manually added the actual changes directly in the master.