Remove examples that suggest getting type from `field.constructor.name`
It is not safe/stable to rely on `.constructor.name` in general because
minification will mangle the names of constructor functions (to make
them as short as possible).

This has confused several folks before, as seen in these issues:
https://github.com/Hopding/pdf-lib/issues/736
https://github.com/Hopding/pdf-lib/issues/755
https://github.com/Hopding/pdf-lib/issues/933
https://github.com/Hopding/pdf-lib/issues/1089

As such, I think we should remove these examples from the codebase.